fff-web: Fix two bugs in wifiscan.html
Bug 1: If devices with hidden SSID are present, the matched
SSID in the awk contains the newline (="unknown\n\t"). This
destroys the table after the sort.

Bug 2: If SSIDs contain ampersands (yes, people do that),
they are now converted to &amp;.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

Reviewed-by: Tim Niemeyer <tim@tn-x.org>

----

Changes in v2:
- none

Changes in v3:
- Use Tim's preferred regex

Changes in v4:
None