fff-config/fff-hoods: Provide a setting to use custom SSIDs
Private patch.

This provides a possibility to set a custom ID for the AP and
Mesh WiFi interfaces. The patch includes a script to set/reset
those correctly.

With this, a user can e.g. use an SSID minikaff.freifunk where
he would otherwise be forced to use remotecity.freifunk, but
without having to set up his own gateway.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>