fff-sysupgrade: Remove dependency from fff-hoods
Sysupgrade implements a lib file from fff-hoods to look for the
hood file and read the upgrade path from it.
If fff-hoods is not desired, this poses a problem. Thus, we
move the lib file to fff-config, as its primary purpose is
actually providing the paths to the hood files (=config).

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>