fff-network/fff-hoods: Calculate fdff IPs based on uci
If the mac is read from /sys/class/net/${iface}/address, some
devices (WA860RE, Picostation) will not set the fdff addresses.

This can be fixed by using the uci value instead of the br-mesh
device, as uci is available instantly.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

Tested-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

---

I don't believe it myself, but it seems to be this tiny change
making the difference.