fff-web: Label Freifunk routers individually in wifiscan.html
This patch enables labelling of routers in the WebUI's WiFi
scan page, so their hostname is displayed instead of
freifunk.franken.net. The evaluation is performed based on
a WifiAnalyzer style node file /tmp/wifinodelist.

In the WebUI, this file may be provided by file upload or by
downloading from the Monitoring API (if the router has
internet access).

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>