fff-web: Label Freifunk routers individually in wifiscan.html
This patch enables labelling of routers in the WebUI's WiFi
scan page, so their hostname is displayed instead of
freifunk.franken.net. The evaluation is performed based on
a WifiAnalyzer style node file /tmp/wifinodelist.

In the WebUI, this file may be provided by file upload or by
downloading from the Monitoring API (if the router has
internet access).

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

Tested-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

---

Changes in v2:
none

Changes in v3:
- Put to the end of the patchset
- Use wifianalall API (hood-independent) as this is now available
  and eves the file for all hoods has only < 200 kB
- Reduced upload file size to 1MB