fff-hoods: Restart alfred after keyxchange file update
This is a workaround to enable alfred on routers which do not see
a configap during initial boot.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

---

Although this might not be popular, it should fix the problem
perfectly. Since I do not think we solve the underlying issue
shortly, I think this patch should help in the meantime.