init.d/fff-hoods: Move call of configurehood to init.d script
This ensures that configurehood is executed AFTER the LEDS are
set up, but BEFORE alfred.

If a configap is reachable, we are thus able to set up w2mesh
before alfred tries to access br-mesh.

As a side effect, we move the call of configurehood to the
fff-hoods package, where it belongs.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>
Tested-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

---

Changes in v2:
none

Changes in v3:
- Changed START=99 to START=98