configurehood: Check for timestamp of hood file over LAN
This patch does no more take the first hood file it finds, but
compares the timestamps provided.

Additionally, we again started to check for the gateway at the
end, but only take the file if its newer and FROM THE SAME HOOD.

Thus, we enable updates, but prohobit switching from a cable
hood to a WiFi peer. (Bug # 68)

Fixes #95.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>