configurenetwork: Fix WAN assignment for special cases
When rewriting the script, the behavior has been altered
unintentionally: If a device has separate eth0 and eth1, but the
WAN port is inside the switch, the port will not work anymore.

This has been fixed, and while at, I added comments and also
fixed the call of enableAutoConf for those cases.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>