fff-config: Split uci-defaults script into early and late part
So far, the 98-configure-fff script did both the setup of the
/etc/config/fff file (if it is was not there) and the application
of settings from it.

In this script, the setup part is moved to a much earlier stage,
so /etc/config/fff is initialized when we need it in other scripts.

While at it, remove error suppression for uci set commands.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>