fff-network/-support/-web: Nice setup of special-port devices
This patch:
- provides scripts to easily change port configuration on the
  "special" devices
- provides a file to store upgrade-safe script lines run at
  first boot (to be used for port adjustments)
- fixes the port page in WebUI
- fixes show_info

With this patch, instead of specifying data in network.config
and then have a script to deal with it, we define a script
/etc/firstbootfff where COMMANDS can be stored. Thus, one can
just make switch ports permanent by putting the uci commands there.

To keep the convenient setup of one-port (and other special)
devices, we implement script, e.g.
setoneport "WAN"
will set the port of a one-port device to WAN mode. Those scripts
are used in the WebUI and may also be implemented manually in the
/etc/firstbootfff file.

This patch misses the selection of the current port mode in the
WebUI. It would have been to much additional code.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>