fff-network: Migrate old settings
This patch will migrate the custom settings from network.config
to the new configuration system.

The uci-default script only writes to /etc/firstbootfff, as this
is evaluated later in the uci-defaults run; it does not directly
apply any settings.
By writing to /etc/firstbootfff, the settings are still kept in
an upgrade-safe way.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

---

Changes in v3:
- Added this patch

This is now the correct version of this patch. Note that this is
completely untested so far.