packages: Update Makefile versioning scheme
With this latest change, the PKG_VERSION is refering to the
version of a non-local package (incl. official firmware).
If the package is unchanged, PKG_RELEASE will be 0.
If the package is changed, PKG_RELEASE will be increased until
a newer external version is pulled in. After that, PKG_RELEASE
will be reset to 0 or 1 depending on whether local changes remain.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>