fff-network: Use existing network config from OpenWrt
This rearranges the initial network config in a way that it is
build into the existing config from OpenWrt. The only change
which requires user interaction will be the port assignment
between BATMAN and CLIENT.

If a new device is added without any adjustments, it will now
work with ports configured as done in OpenWrt, i.e. switch or
one-port assigned to client.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>