(GW) fff-gateway: Do not reset whole switch in configuregateway
Instead of (again) setting up the whole switch in
/etc/config/gateway, after this patch only "changes" will need to
be set up there.

For adding or updating a VLAN:

config add_vlan
	option vlan '11'
	option ports '0t 4'

For removing a VLAN (e.g. if the VLAN ID is changed):

config del_vlan
	option vlan '2'

The config blocks can be named, but the name is not evaluated.

If no change is desired (e.g. to work with default config) or if
config has already been updated otherwise (e.g. in
/etc/firstbootfff), none of the blocks is necessary.

However, one will still have to assign the VLAN IDs to the client,
batman, etc. blocks, e.g.

config batman
        option vlan '3'

In addition to the new setup rules, the CPU port now needs to be
specified by the user. This removes the old code to set it.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>