fff-network: Tidy up MAC address setup
This adjusts the sources for br-mesh and ethmesh MAC addresses:

Ambiguous addresses for mesh are modified to use phy0, as this
will yield a large default case.

Ethmesh setup is changed to calculate the MAC address according
to how device addresses are set. Thus, we do not depend on
correctly choosing the source device, but still only use
addresses which are actually assigned to the device.

The following devices do not need an ethmesh MAC address, since
their eth MAC addresses do not overlap with WiFi, so they are
removed from the assignment:
- Archer C7 v2
- Archer C25 v1
- Archer C60 v1/v2
- TL-WR740N/ND v4
- TL-WR741N/ND v2/v4
- TL-WR841N/ND v7
- All Ubiquiti devices

The ethmesh MAC address for the TL-WDR4900 v1 changes, the
WAN MAC address is used.

The two files for MAC address setup are merged into one.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>