fff-network/others: Convert ethmesh to bridge
This will allow to easily set BATMAN mode to ethernet multiple
interfaces. While one can in principle do this by adding proto
batman to multiple interfaces, functions relying on a single
interface for ethernet mesh (e.g. getEthernetHoodfile) will
break.

This requires to set ethmesh MAC differently, as we do need
different individual MAC addresses for each device in ethmesh.
This is done by adding a device to the network config, as OpenWrt
does it for WAN.
Note that this only works for setting ethmesh MAC on switch-based
and one-port devices, but this is sufficient at the moment.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>