(GW) fff-gateway: Add option to set WAN iface/vlan in config
While setting client and batman is required, the wan entry is
only required if wan binding should change:

config wan
	option vlan '12'

config wan
	option iface 'eth0'

If the block is missing, wan will be left untouched. If it should
be removed, this has to be done manually, as it will be a very
rare case.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>