bsp: Repair and align .config files for different targets
This will align options in the .config files to be as similar as
possible among different targets. This will also switch to
per-target root FS for targets with only one device.

Packages which are just selected with "y" are moved to the fff
Makefile. Definitions are put into blocks to enhance readability.

Remove CONFIG_STRIP_KERNEL_EXPORTS for all non-tiny targets.

Remove explicit USB packages selection, as this will be done
automatically by devices that requiring them.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>