gre: Introduce 'nohostroute' option
It is not always necessary to add a host route for the gre peer address.

This introduces a new config option 'nohostroute' (similar to the
option introduced for wireguard in d8e2e19) to allow to disable
the creation of those routes explicitely.

This is a backport from OpenWrt master.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>