ath79: Add support for Joy-IT JIT-OR750i
This adds the Joy-IT JIT-OR750i device based on an experimental
OpenWrt patch by David Bauer.

The device is added for and maintained in ath79 to save us from
having to change it again in the near future.

Label MAC address is on WAN.

WAN   art 0x0   label
LAN   art 0x6   label + 1
phy1 (2.4 GHz)  label + 2
phy0 (5 GHz)    label + 3

For ath79, WAN is eth1 and LAN is eth0. For ar71xx, the assignment
is vice-versa.

This creates a new bsp file for ath79 devices to be used with
19.07-based firmware (ath79-back.bsp).

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>