treewide: Make device names more systematic
This removes the forceful rename of ar71xx and/or ath79 devices to
get them have the same names. Instead, all boardnames except ar71xx
are used unchanged, and ar71xx itself keeps the legacy names with
rewrites as in official firmware.

The only exception from this is done for Ubiquiti devices, where
the "ubnt-" prefix is removed, so names correspond to the original
ar71xx style.

This introduces some additional changes on the way:
- The PORTORDER for ar71xx is removed, as this device is already
  treated as two-port.
- CPE210/CPE510 have different setup on ar71xx and ath79. For them
  the different names are exploited to have different setup.
  Attention: So far, no migration script is present for manual
  ethernet config!
- pico-m and power-m-xw are only set up for ar71xx at the moment.
- Remove board identification for ath79, since this target
  identifies by choice of image

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>