fff-sysupgrade: remove target and subtarget from filename
This simplifies and shortens filenames quite significantly.

A rewrite script will be installed on the update servers
to allow updating routers with older firmwares.

Signed-off-by: Fabian Bläse <fabian@blaese.de>
Reviewed-by: Christian Dresel <fff@chrisi01.de>

ported from: 3e1f6e167d4b ("Remove target and subtarget from filename")

This ports the changes to sysupgrade.sh/fff-upgrade.sh, as this
script is meant to download official firmware only.

No change is made to buildscript, as this firmware does not intend
to remove the target name.