checkpatch.pl: introduce OpenWrt script to check patches
This copies the OpenWrt script used to check patches before submission.
This is meant as a help and guidance for both submitters and reviewers,
as it will spot many formal and stylistic errors.

The script is used on one or several patches formatted for submission
to the mailing list (i.e. created by "git format-patch"):

./checkpatch.pl abc.patch def.patch

This patches just copies the unaltered script from the OpenWrt trunk.
It will be adjusted to our needs in a subsequent patch.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>