checkpatch.pl: adjust to be used with Freifunk Franken firmware
While checkpatch.pl is meant to be run in OpenWrt root directory,
is already provides options for when this is not the case.

This patch alters these options so out-of-tree is default, and
in-tree variant cannot be enabled as it's useless for us. As we
do not have to check kernel patches etc. for our firmware directly
this does not pose a problem.

While one technically could remove the unused parts of the code,
this patch keeps them and only changes the relevant variables.
This should provide easy updates in case of upstream changes.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>