ramips/mt76x8: add support for TP-Link Archer C50 v4
Add support for the TP-Link Archer C50 v4, a low-cost mt7628-based
dual-band router.

Since this is the first device on mt76x8 (sub)target in our firmware,
add the target as well.

The device name will be updated from tplink,c50-v4 to
tplink,archer-c50-v4 in the next stable branch of OpenWrt, so
make the firmware use that string already.

Label MAC address is on ethernet and 2.4 GHz WiFi.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>