treewide: move devices from ar71xx to ath79
This patch removes device support for ar71xx devices that are supported
in ath79. Building them with ath79 will be the new default.

The only devices remaining in ar71xx will be the following, as there
exists no ath79 support for them in master or backporting is not
desirable at the moment:
- cpe210-v1
- cpe510-v1

The ar71xx-tiny subtarget will be removed as there is no device left.

The following device/board is dropped entirely as it essentially was a hack
built based on another device and official OpenWrt support is not expected:
- power-m-xw

For the tl-wr741nd-v2, in ath79 the tplink_tl-wr741-v1 image is used.

As a consequence, this drops/renames all affected board identifiers specific
to ar71xx. Furthermore, the obsolete file ath79-back.bsp is removed.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>