fff-sysupgrade: always preserve /etc/config/gateway
So far, the file /etc/config/gateway has only been preserved through
upgrade when the GW variant was used as source.

As this file does not interfere with the node variant, though,
keeping it also there will be beneficial in case of flashing the
wrong variant by mistake, but won't have any negative consequences.

Thus, this patch moves the inclusion into sysupgrade.conf from
fff-gateway back to fff-sysupgrade. This also matches the behavior
in official FW.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>