fff-gwinfo: Make curl silent
As we now log output of cron scripts to syslog, the debug
output of curl spams the rather short-lived syslog. As this
debug output is unessesary most of the time, it now is disabled.

Instead, curl's silent option is used, together with '-S', which
reenables error output.

Based on the similar patch for fff-alfred-monitoring-proxy.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>