build_patches: backport fix for WAN disconnect on ath79
The switch driver on ath79 leads to occasional (sometimes even frequent)
dis- and reconnects of the WAN port. This has finally been resolved
by a workaround applied to OpenWrt master now. The problem also existed
with ar71xx, but was fixed a long time ago already.

This cherry-picks the relevant patch to 19.07 and adds it to our
firmware (it will probably be included in 19.07.5).

The problem as well as the resolution by the patch has been reproduced
on device (with my FFF firmware and with 19.07.4 OpenWrt).

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>