fff-network: set random MAC address for ethmesh bridge
By design, batman-adv's bridge-loop-avoidance (BLA) just uses the
address of the first member interface of bat0 [1] for the packets
it sends. However, this will create collisions with the members
of br-mesh and result in annoying warnings that spam kernel (and
system) log:

[  179.939430] br-mesh: received packet on bat0 with own address as
               source address (addr:fa:1a:67:xx:xx:fb, vlan:0)

Being created by the BLA packets itself, one possible option would
obviously be to just disable BLA.

Another option - with keeping BLA enabled - will be to choose a
random MAC address for the ethmesh bridge device, which happens
to be the primary bat0 member device. Though this brings the
disadvantage of not having a known MAC address for link-local on
ethernet mesh, this should be easy to address by using a multicast
ping.

The random MAC address is created via uci-defaults, so it will
persistent across reboots, but be reset during upgrade.

In addition to the change itself, this backports a few functions
from OpenWrt master that provide the macaddr_random function.

[1] https://www.open-mesh.org/issues/421

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>